feat: changed bottom nav bar setup
