+++ app/devtools_options.yaml
@@ -1,0 +1,1 @@
+extensions:
