+++ app/lib/presentation/screens/home.dart
@@ -1,12 +1,7 @@
-import 'package:bankimoon/data/Models/accounts.dart';
-import 'package:bankimoon/data/cubit/accounts_cubit.dart';
-import 'package:bankimoon/presentation/widgets/account_list_widget.dart';
-import 'package:bankimoon/presentation/widgets/search_results_widget.dart';
-import 'package:bankimoon/utils/constants.dart';
+import 'package:bankimoon/utils/lists.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:go_router/go_router.dart';
 
-import '../widgets/nav_bar.dart';
 
 class Home extends StatefulWidget {
   const Home({super.key});
@@ -16,93 +11,52 @@ class Home extends StatefulWidget {
 }
 
 class _HomeState extends State<Home> {
-
-  @override
-  void initState() {
-
-    WidgetsFlutterBinding.ensureInitialized().addPostFrameCallback((timeStamp) {
-      BlocProvider.of<AccountsCubit>(context).stream.listen((state) {
-        if (state is ErrorState) {
-          ScaffoldMessenger.of(context).showSnackBar(
-            SnackBar(
-              content: Text(state.message),
-              backgroundColor: Colors.red,
-            ),
-          );
-        }
-      });
-
-      BlocProvider.of<AccountsCubit>(context).getUserAccounts();
-    });
-
-    super.initState();
-  }
+  int _currentIndex = 0;
 
   @override
   Widget build(BuildContext context) {
-    final size = MediaQuery.of(context).size;
-
     return Scaffold(
-      appBar: AppBar(
-        title: const Text(
-          "Bankimoon",
-          style: titleStyles,
-        ),
-        centerTitle: true,
-        leading: const Icon(
-          Icons.ac_unit,
-          color: Color.fromARGB(255, 15, 91, 254),
-        ),
-        bottom: PreferredSize(
-          preferredSize: const Size.fromHeight(50.0),
-          child: Padding(
-            padding: const EdgeInsets.all(8.0),
-            child: SizedBox(
-              height: 45.0,
-              child: SearchBar(
-                leading: const Icon(Icons.search),
-                hintText: "Search",
-                onChanged: (value) {
-                  if (value.isNotEmpty) {
-                    BlocProvider.of<AccountsCubit>(context)
-                        .searchAccount(value.toString());
-                  } else {
-                    BlocProvider.of<AccountsCubit>(context).getUserAccounts();
-                  }
-                },
-              ),
-              //width: 300.0,
-            ),
-          ),
+      body: screens[_currentIndex],
+      floatingActionButton: FloatingActionButton.extended(
+        onPressed: () {
+          context.push("/add-account");
+        },
+        label: const Text("New "),
+        icon: const Icon(
+          Icons.add,
         ),
       ),
-      body: BlocBuilder<AccountsCubit, AccountsState>(
-        builder: (context, state) {
-          List<Account> accounts = [];
-
-          if (state is AccountsFetched) {
-            accounts = state.accounts;
-          } else if (state is AccountSearchResults) {
-            return SearchResultsWidget(accounts: state.accounts);
-          } else if (state is FetchingAccounts || state is DeletingAccount) {
-            return SizedBox(
-              height: size.height,
-              width: size.width,
-              child: const Center(
-                child: CircularProgressIndicator(
-                  color: Color.fromARGB(255, 15, 91, 254),
-                ),
-              ),
-            );
-          }
-
-          return AccountListWidget(accounts: BlocProvider.of<AccountsCubit>(context).accounts, onDismissed: (index){
-              BlocProvider.of<AccountsCubit>(context)
-                  .deleteAccount(accounts[index].id!);
+      bottomNavigationBar: BottomNavigationBar(
+        onTap: (int index) {
+          setState(() {
+            _currentIndex = index;
           });
         },
+        type: BottomNavigationBarType.fixed,
+        unselectedItemColor: const Color.fromARGB(255, 144, 168, 236),
+        selectedItemColor: Colors.white,
+        currentIndex: _currentIndex,
+        backgroundColor: const Color.fromARGB(255, 15, 91, 254),
+        items: const [
+          BottomNavigationBarItem(
+            icon: Icon(
+              Icons.home,
+            ),
+            label: "Home",
+          ),
+          BottomNavigationBarItem(
+              icon: Icon(
+                Icons.favorite,
+              ),
+              label: "Favorites"),
+          BottomNavigationBarItem(
+            icon: Icon(
+              Icons.info,
+            ),
+            label: "About",
+          )
+        ],
       ),
-      bottomNavigationBar: const BottomNavBar(),
     );
   }
 }
