+++ app/lib/utils/router.dart
@@ -1,6 +1,5 @@
 import 'package:bankimoon/data/cubit/accounts_cubit.dart';
 import 'package:bankimoon/presentation/screens/add_account.dart';
-import 'package:bankimoon/presentation/screens/favourites.dart';
 import 'package:bankimoon/presentation/screens/splash_screen.dart';
 import 'package:bankimoon/presentation/screens/home.dart';
 import 'package:bankimoon/presentation/screens/about.dart';
@@ -24,14 +23,14 @@ final GoRouter router = GoRouter(
         child: const Home(),
       ),
     ),
-    GoRoute(
-      name: "favourites",
-      path: favouritePage,
-      builder: (context, state) => BlocProvider(
-        create: (context) => AccountsCubit(),
-        child: const Favourite(),
-      ),
-    ),
+    // GoRoute(
+    //   name: "favourites",
+    //   path: favouritePage,
+    //   builder: (context, state) => BlocProvider(
+    //     create: (context) => AccountsCubit(),
+    //     child: const Favourite(),
+    //   ),
+    // ),
     GoRoute(
       name: "add-account",
       path: addAccount,
